+++ tests/testthat/test_signif2.R
@@ -6,12 +6,12 @@ test_that("Several tests",{
     expect_equal(signif2(x,3),
                  c("0.000124","1130000","1.10","22.0","10.0","1.00")
                  )
-    expect_warning(signif2(x,3,add=FALSE))
+##    expect_warning(signif2(x,3,add=FALSE))
     expect_equal(signif2(c(.2,11.84),2),
-                 c("0.2","12")
+                 c("0.20","12")
                  )
     expect_error(signif2(3205,-1))
     expect_equal(signif2(0,1),"0")
-    expect_equal(signif2(0,3),"0.00")
+    expect_equal(signif2(0,3),"0")
 
 })